aircraft-json-collector/requirements.txt
requests==2.21.0
google-cloud-pubsub==0.39.1
nose==1.3.7